Shifted and renamed the proof and credential models into a new location
