+++ src/Streetcred.Sdk/Contracts/IProofService.cs
@@ -3,9 +3,10 @@ using System.Threading.Tasks;
 using Hyperledger.Indy.PoolApi;
 using Hyperledger.Indy.WalletApi;
 using Streetcred.Sdk.Messages.Proofs;
+using Streetcred.Sdk.Models.Credentials;
+using Streetcred.Sdk.Models.Proofs;
 using Streetcred.Sdk.Models.Records;
 using Streetcred.Sdk.Models.Records.Search;
-using Streetcred.Sdk.Models.Wallets;
 
 namespace Streetcred.Sdk.Contracts
 {
@@ -21,7 +22,7 @@ namespace Streetcred.Sdk.Contracts
         /// <param name="wallet">Wallet.</param>
         /// <param name="connectionId">Connection identifier of who the proof request will be sent to.</param>
         /// <param name="proofRequest">Proof request object describing the proof request.</param>
-        Task SendProofRequestAsync(Wallet wallet, string connectionId, ProofRequestObject proofRequest);
+        Task SendProofRequestAsync(Wallet wallet, string connectionId, ProofRequest proofRequest);
 
         /// <summary>
         /// Sends a proof request
@@ -38,9 +39,9 @@ namespace Streetcred.Sdk.Contracts
         /// <returns>The proof request.</returns>
         /// <param name="wallet">Wallet.</param>
         /// <param name="connectionId">Connection identifier of who the proof request will be sent to.</param>
-        /// <param name="proofRequestObject">An enumeration of attribute we wish the prover to disclose.</param>
+        /// <param name="proofRequest">An enumeration of attribute we wish the prover to disclose.</param>
         Task<ProofRequestMessage> CreateProofRequestAsync(Wallet wallet, string connectionId,
-            ProofRequestObject proofRequestObject);
+            ProofRequest proofRequest);
 
         /// <summary>
         /// Creates a proof request
@@ -78,7 +79,7 @@ namespace Streetcred.Sdk.Contracts
         /// The proof.
         /// </returns>
         Task<ProofMessage> CreateProofAsync(Wallet wallet, Pool pool, string proofRequestId,
-            RequestedCredentialsDto requestedCredentials);
+            RequestedCredentials requestedCredentials);
 
         /// <summary>
         /// Accepts a proof request by generating a proof and sending it to the requestor
@@ -91,7 +92,7 @@ namespace Streetcred.Sdk.Contracts
         /// The proof.
         /// </returns>
         Task AcceptProofRequestAsync(Wallet wallet, Pool pool, string proofRequestId,
-            RequestedCredentialsDto requestedCredentials);
+            RequestedCredentials requestedCredentials);
 
         /// <summary>
         /// Rejects a proof request
@@ -133,13 +134,13 @@ namespace Streetcred.Sdk.Contracts
         /// Lists the credentials available for the given proof request.
         /// </summary>
         /// <param name="wallet">The wallet.</param>
-        /// <param name="proofRequestObject">The proof request object.</param>
+        /// <param name="proofRequest">The proof request object.</param>
         /// <param name="attributeReferent">The attribute referent.</param>
         /// <returns>
-        /// A collection of <see cref="CredentialObject" /> that are available
+        /// A collection of <see cref="CredentialInfo" /> that are available
         /// for building a proof for the given proof request
         /// </returns>
-        Task<List<CredentialInfo>> ListCredentialsForProofRequestAsync(Wallet wallet,
-            ProofRequestObject proofRequestObject, string attributeReferent);
+        Task<List<Credential>> ListCredentialsForProofRequestAsync(Wallet wallet,
+            ProofRequest proofRequest, string attributeReferent);
     }
 }
