+++ test/Streetcred.Sdk.Tests/ProofTests.cs
@@ -13,8 +13,8 @@ using Streetcred.Sdk.Contracts;
 using Streetcred.Sdk.Messages;
 using Streetcred.Sdk.Messages.Proofs;
 using Streetcred.Sdk.Models;
+using Streetcred.Sdk.Models.Proofs;
 using Streetcred.Sdk.Models.Records;
-using Streetcred.Sdk.Models.Wallets;
 using Streetcred.Sdk.Runtime;
 using Xunit;
 
@@ -166,7 +166,7 @@ namespace Streetcred.Sdk.Tests
 
             // Verifier sends a proof request to prover
             {
-                var proofRequestObject = new ProofRequestObject
+                var proofRequestObject = new ProofRequest
                 {
                     Name = "ProofReq",
                     Version = "1.0",
@@ -192,9 +192,9 @@ namespace Streetcred.Sdk.Tests
                 var holderProofRequestId = await _proofService.ProcessProofRequestAsync(_holderWallet, proofRequest);
                 var holderProofRecord = await _proofService.GetAsync(_holderWallet, holderProofRequestId);
                 var holderProofObject =
-                    JsonConvert.DeserializeObject<ProofRequestObject>(holderProofRecord.RequestJson);
+                    JsonConvert.DeserializeObject<ProofRequest>(holderProofRecord.RequestJson);
 
-                var requestedCredentials = new RequestedCredentialsDto();
+                var requestedCredentials = new RequestedCredentials();
                 foreach (var requestedAttribute in holderProofObject.RequestedAttributes)
                 {
                     var credentials =
@@ -202,9 +202,9 @@ namespace Streetcred.Sdk.Tests
                             requestedAttribute.Key);
 
                     requestedCredentials.RequestedAttributes.Add(requestedAttribute.Key,
-                        new RequestedAttributeDto
+                        new RequestedAttribute
                         {
-                            CredentialId = credentials.First().CredentialObject.Referent,
+                            CredentialId = credentials.First().CredentialInfo.Referent,
                             Revealed = true,
                             Timestamp = DateTimeOffset.UtcNow.ToUnixTimeSeconds()
                         });
@@ -217,9 +217,9 @@ namespace Streetcred.Sdk.Tests
                             requestedAttribute.Key);
 
                     requestedCredentials.RequestedPredicates.Add(requestedAttribute.Key,
-                        new RequestedAttributeDto
+                        new RequestedAttribute
                         {
-                            CredentialId = credentials.First().CredentialObject.Referent,
+                            CredentialId = credentials.First().CredentialInfo.Referent,
                             Revealed = true
                         });
                 }
